Enable wmt_dbg functionality
These patches make it possible to read memory from the WiFi controller's
address space. Functions executed by `wmt_dev_dbg_write` can be accessed
as follows:

echo b 00000000 0 > /proc/driver/wmt_dbg

Where "b" is the hexadecimal index of the function you want to run in
`wmt_dev_dbg_func[]` and "00000000" and "0" are the first and second
arguments to the function (`par2` and `par3` in the code).

In this case, 0xb is `wmt_dbg_reg_read`, so "00000000" is the address of
the register/dword you want to read (zero-padding not required), and "0"
is supposedly a mask value, but it doesn't seem to actually do anything.

The result of the read can be viewed by grepping dmesg for
"wmt_dbg_reg_read".