Fix MT6737M JTAG register-setting preloader patch
This SoC is software-compatible with the MT6735, so the GPIO mode
register should be exactly the same. I'm still not sure why I haven't
been able to get JTAG working, but this change has been sitting around
uncommitted for a while now and I figure there's at least no harm in
doing that now.