Set timeout offset for the official Orange Pi 4G-IoT preloader
The previous offset only worked with the custom version of the preloaer
I built myself, but this new offset works with the one that can be
downloaded from the Orange Pi website. At some point, these hard-coded
offsets should be replaced with a hash-based lookup of patches, so
multiple binaries can be supported. I should also probably write a
virtual timer peripheral that just runs slower so I don't have to patch
the timeout function on every binary.