Add a small program that can write an arbitrary sequence of registers
This code can be used to set the SRAM variables necessary to disable the
memory bounds checks on USB DL mode memory read and write functions.
This enables the USB DL mode host to read and write to memory that would
normally be blocked by the device.

This code is meant to be loaded into a SoC as a certificate and then
executed using the "amonet-kamakiri" method. It takes arguments as an
array of 32-bit little-endian integers appended to the binary at load
time. After it finishes all the memory writes, it returns control to the
BROM so DL mode can resume. Ideally, this code should be
platform-independent, but I've only tested this with an MT6771/MT8183,
so it may or may not work with other devices.