Add register write sequence needed to disable bounds checks on MT6771
Technically, this doesn't disable the bounds checks--it only modifies
the permission table so memory read and write commands can access any
address. Note that writing this specific sequence will prevent other
commands from being able to read from or write to memory (I think),
since the new table does not include any permissions for the other
commands, so if that functionality is desired the bitmasks for the other
commands will need to be added to the table.