Remove NOP sled before PCM code
After some testing, I found that the PCM reset vector is zero, so the
NOP sled isn't needed. I'm not really sure why my code wasn't working
without it before, but it does now, so let's remove it. Also, change the
instructions to register writes so it's easier to tell that it's
working.