Truncate the BROM log if the log length is available
The BROM log length gets set to zero when the log is either dumped out
the UART or read by one of the DL mode "read BROM log" commands
(0xDD/0xDF), so we only limit the the output length when the reported
length is non-zero. When the reported length is zero, we simply print
each character until we either hit the default max length (0x400 bytes,
the length of the BROM log buffer) or encounter a non-printable
character.