Fix licenses for socemu.py and bmo.py
Unicorn is based on Qemu, which as a whole is licensed under only
version 2 of the GPL. Since socemu.py depends on Unicorn, socemu.py
needs to be licensed under a GPLv2-compatible license. socemu.py also
depends on bmo.py, so to ensure that the combined work can be
distributed under the GPLv2 bmo.py also needs to be licensed under a
GPLv2-compatible license.

Fix both of these issues by relicensing both socemu.py and bmo.py under
version 2 or later of the GNU GPL.